test(core): split render/markdown.rs tests by rendered surface
Move the 3240-line inline `#[cfg(test)] mod tests` block out of
`rinkaku-core/src/render/markdown.rs` into
`rinkaku-core/src/render/markdown_tests/`, grouped by which section
of the Markdown output the tests pin.

- `empty_and_ordering` — empty-report short-circuit and the
  Tests / Other-changed / Skipped-files section ordering
- `change_graph_summary` — the "N (changed) symbols in M files — most
  in ..." line, including origin-dependent wording and tie-breaks
- `change_graph_tree` — nesting, pre-order DFS, `(see above)`, cycle
  warnings, ADR 0012 folding
- `definition_body` — per-symbol `### ...` rendering, `Depends on:`,
  fence widening for backtick runs
- `sections_skipped_hotspots_filesize` — Skipped files, Hotspots,
  File size warnings sections, plus the ADR 0028 JSON shape
- `lookup_miss_defenses` — the three defensive
  `SymbolLookup::get() -> None` branches
- `classification_and_removed` — ADR 0014 markers, ` ```diff ` block,
  Removed symbols (with dedup)

`markdown.rs` drops from 3987 to 749 lines. No production body
changed. Test count unchanged (351 rinkaku_core tests still pass).
