+++ rinkaku-core/src/render/markdown.rs
@@ -745,3243 +745,5 @@ fn longest_backtick_run(text: &str) -> Option<usize> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::diff::LineRange;
-    use crate::extract::SymbolKind;
-    use crate::graph::{Edge, Node};
-    use crate::render::report::{
-        FileReport, ReportOrigin, SkipReason, SkippedFile, TestFileSummary,
-    };
-    use crate::render::{OutputFormat, render};
-    use pretty_assertions::assert_eq;
-
-    /// Builds an `ExtractedSymbol` for rendering tests, with `id` set (the
-    /// graph-building pipeline stage this module assumes already ran) and
-    /// every other field defaulted to something inert unless overridden via
-    /// struct-update syntax at the call site.
-    fn symbol(id: &str, name: &str, kind: SymbolKind, signature: &str) -> ExtractedSymbol {
-        ExtractedSymbol {
-            id: id.to_string(),
-            name: name.to_string(),
-            kind,
-            signature: signature.to_string(),
-            range: LineRange { start: 1, end: 1 },
-            container: None,
-            referenced_names: vec![],
-            dependencies: vec![],
-            omitted_dependency_matches: 0,
-            is_test: false,
-            classification: None,
-            previous_signature: None,
-        }
-    }
-
-    fn node(id: &str, path: &str, name: &str) -> Node {
-        Node {
-            id: id.to_string(),
-            path: path.to_string(),
-            name: name.to_string(),
-        }
-    }
-
-    #[test]
-    fn should_render_empty_markdown_when_report_has_no_files_and_no_skips() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "".to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    // Regression test: a pure rename (or mode-change-only diff) is reported
-    // as a `FileReport` with an empty `symbols` list (see
-    // `pipeline::analyze_diff`'s doc comment) rather than a `SkippedFile` —
-    // the file *was* looked at, it just had no symbol-level changes. Before
-    // this fix, such a file was silently dropped from Markdown output
-    // entirely (the empty-output guard fired because `graph.nodes` and
-    // `skipped` were both empty, even though `files` was not), which is a
-    // regression from the pre-ADR-0008 renderer that always emitted a `##
-    // {path}` heading for every entry in `report.files`.
-    #[test]
-    fn should_list_file_with_no_symbols_under_other_changed_files_when_report_has_no_graph_nodes() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/new_name.rs".to_string(),
-                symbols: vec![],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Other changed files
-
-- src/new_name.rs
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_list_file_with_no_symbols_after_definitions_when_report_has_graph_nodes_too() {
-        // A diff with one file that has a changed symbol (feeds the
-        // "Change graph"/"Definitions" sections) alongside a pure-rename
-        // file with no symbols at all — the pure rename must still show up,
-        // in its own section after "Definitions".
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![
-                FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![symbol(
-                        "src/lib.rs::foo",
-                        "foo",
-                        SymbolKind::Function,
-                        "fn foo()",
-                    )],
-                },
-                FileReport {
-                    path: "src/new_name.rs".to_string(),
-                    symbols: vec![],
-                },
-            ],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-## Other changed files
-
-- src/new_name.rs
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_other_changed_files_before_skipped_files_when_report_has_both() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/new_name.rs".to_string(),
-                symbols: vec![],
-            }],
-            skipped: vec![SkippedFile {
-                path: "assets/logo.png".to_string(),
-                reason: SkipReason::Binary,
-            }],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Other changed files
-
-- src/new_name.rs
-
-## Skipped files
-
-- assets/logo.png (binary)
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_tests_section_with_singular_symbol_noun_when_count_is_one() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![crate::render::TestFileSummary {
-                path: "src/lib.rs".to_string(),
-                symbol_count: 1,
-            }],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Tests
-
-- src/lib.rs: 1 changed test symbol
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_tests_section_with_plural_symbols_noun_when_count_is_greater_than_one() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![crate::render::TestFileSummary {
-                path: "src/lib.rs".to_string(),
-                symbol_count: 3,
-            }],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Tests
-
-- src/lib.rs: 3 changed test symbols
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_tests_section_between_definitions_and_other_changed_files_when_report_has_all_sections()
-     {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/lib.rs::foo",
-                    "foo",
-                    SymbolKind::Function,
-                    "fn foo()",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![crate::render::TestFileSummary {
-                path: "src/lib.rs".to_string(),
-                symbol_count: 2,
-            }],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-## Tests
-
-- src/lib.rs: 2 changed test symbols
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    // Regression test: a `Generated` skip entry must not appear in Markdown
-    // output at all (not even under "Skipped files") — `.gitattributes`
-    // already marks these files as uninteresting to diff-review, so
-    // Markdown output (meant for humans/LLMs skimming a change) drops them
-    // silently rather than listing them as something rinkaku "didn't look
-    // at". They stay visible in JSON (see
-    // `should_keep_generated_entry_in_json_output` below) for machine
-    // consumers that want the full picture.
-    #[test]
-    fn should_omit_generated_skip_entry_from_markdown_output() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![SkippedFile {
-                path: "Cargo.lock".to_string(),
-                reason: SkipReason::Generated,
-            }],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "".to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    // Sibling case: when a `Generated` entry is mixed with other skip
-    // reasons, only the generated one is dropped from Markdown — the
-    // section itself still renders for the remaining, non-generated skips.
-    #[test]
-    fn should_omit_only_generated_entries_when_skipped_has_other_reasons_too() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![
-                SkippedFile {
-                    path: "Cargo.lock".to_string(),
-                    reason: SkipReason::Generated,
-                },
-                SkippedFile {
-                    path: "assets/logo.png".to_string(),
-                    reason: SkipReason::Binary,
-                },
-            ],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Skipped files
-
-- assets/logo.png (binary)
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-    #[test]
-    fn should_render_change_graph_and_definitions_when_report_has_one_symbol() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/lib.rs::foo",
-                    "foo",
-                    SymbolKind::Function,
-                    "fn foo(a: i32) -> i32",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo(a: i32) -> i32
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_summary_with_hotspot_when_report_has_multiple_symbols_and_files() {
-        // 5 nodes across store/items.go (3, the hotspot) and store/db.go (2)
-        // — pins the plural "changed symbols"/"files" wording together with
-        // the "— most in ..." suffix and its count.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("store/items.go::A", "store/items.go", "A"),
-                    node("store/items.go::B", "store/items.go", "B"),
-                    node("store/items.go::C", "store/items.go", "C"),
-                    node("store/db.go::D", "store/db.go", "D"),
-                    node("store/db.go::E", "store/db.go", "E"),
-                ],
-                edges: vec![],
-                roots: vec![
-                    "store/items.go::A".to_string(),
-                    "store/items.go::B".to_string(),
-                    "store/items.go::C".to_string(),
-                    "store/db.go::D".to_string(),
-                    "store/db.go::E".to_string(),
-                ],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-5 changed symbols in 2 files — most in store/items.go (3)
-
-
-## Definitions
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    // ADR 0017: a whole-repo outline has no diff, so "Change graph"/
-    // "changed symbols" would misdescribe it — this pins the alternate
-    // heading and noun `ReportOrigin::RepoOutline` selects, using the same
-    // multi-file/hotspot shape as
-    // `should_render_summary_with_hotspot_when_report_has_multiple_symbols_and_files`
-    // so the two tests differ only in `origin` and its wording.
-    #[test]
-    fn should_render_repository_graph_heading_and_drop_changed_wording_when_origin_is_repo_outline()
-    {
-        let report = Report {
-            origin: ReportOrigin::RepoOutline,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("store/items.go::A", "store/items.go", "A"),
-                    node("store/items.go::B", "store/items.go", "B"),
-                    node("store/items.go::C", "store/items.go", "C"),
-                    node("store/db.go::D", "store/db.go", "D"),
-                    node("store/db.go::E", "store/db.go", "E"),
-                ],
-                edges: vec![],
-                roots: vec![
-                    "store/items.go::A".to_string(),
-                    "store/items.go::B".to_string(),
-                    "store/items.go::C".to_string(),
-                    "store/db.go::D".to_string(),
-                    "store/db.go::E".to_string(),
-                ],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Repository graph
-
-5 symbols in 2 files — most in store/items.go (3)
-
-
-## Definitions
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_omit_hotspot_suffix_when_all_symbols_are_in_one_file() {
-        // Every node lives in the same file, so naming "the file with the
-        // most nodes" would be redundant — the suffix must be dropped
-        // entirely, not degenerate into e.g. "(2)".
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::a", "src/lib.rs", "a"),
-                    node("src/lib.rs::b", "src/lib.rs", "b"),
-                ],
-                edges: vec![],
-                roots: vec!["src/lib.rs::a".to_string(), "src/lib.rs::b".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-2 changed symbols in 1 file
-
-
-## Definitions
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_break_hotspot_tie_by_first_seen_path_order_when_counts_are_equal() {
-        // b.rs and a.rs both have 2 nodes each; b.rs's node appears first in
-        // `graph.nodes`, so it must win the tie over a.rs despite sorting
-        // after it alphabetically — the tie-break is source order, not a
-        // path-string comparison.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("b.rs::x", "b.rs", "x"),
-                    node("a.rs::y", "a.rs", "y"),
-                    node("b.rs::z", "b.rs", "z"),
-                    node("a.rs::w", "a.rs", "w"),
-                ],
-                edges: vec![],
-                roots: vec![
-                    "b.rs::x".to_string(),
-                    "a.rs::y".to_string(),
-                    "b.rs::z".to_string(),
-                    "a.rs::w".to_string(),
-                ],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-4 changed symbols in 2 files — most in b.rs (2)
-
-
-## Definitions
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_include_start_line_in_label_when_node_id_is_disambiguated_by_line() {
-        // `graph::collect_nodes` appends `@{start_line}` to a node's id only
-        // when its `(path, name)` pair is not unique in the report (e.g.
-        // two overloaded free functions sharing a name). Without a visible
-        // line number, "Change graph"/"Definitions" would show two
-        // identical-looking `fn foo (src/lib.rs)` entries with no way to
-        // tell them apart.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![
-                    ExtractedSymbol {
-                        range: LineRange { start: 1, end: 3 },
-                        ..symbol(
-                            "src/lib.rs::foo@1",
-                            "foo",
-                            SymbolKind::Function,
-                            "fn foo(a: i32)",
-                        )
-                    },
-                    ExtractedSymbol {
-                        range: LineRange { start: 10, end: 12 },
-                        ..symbol(
-                            "src/lib.rs::foo@10",
-                            "foo",
-                            SymbolKind::Function,
-                            "fn foo(a: i32, b: i32)",
-                        )
-                    },
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::foo@1", "src/lib.rs", "foo"),
-                    node("src/lib.rs::foo@10", "src/lib.rs", "foo"),
-                ],
-                edges: vec![],
-                roots: vec![
-                    "src/lib.rs::foo@1".to_string(),
-                    "src/lib.rs::foo@10".to_string(),
-                ],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-2 changed symbols in 1 file
-
-- fn foo (src/lib.rs:1)
-- fn foo (src/lib.rs:10)
-
-## Definitions
-
-### fn foo (src/lib.rs:1)
-
-```
-fn foo(a: i32)
-```
-
-### fn foo (src/lib.rs:10)
-
-```
-fn foo(a: i32, b: i32)
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_nest_callee_under_caller_in_change_graph_when_symbol_references_another() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/main.rs".to_string(),
-                symbols: vec![
-                    symbol(
-                        "src/main.rs::handle_pr",
-                        "handle_pr",
-                        SymbolKind::Function,
-                        "fn handle_pr(args: PrArgs) -> Result<()>",
-                    ),
-                    symbol(
-                        "src/main.rs::resolve_pr_base_sha",
-                        "resolve_pr_base_sha",
-                        SymbolKind::Function,
-                        "fn resolve_pr_base_sha() -> Result<String>",
-                    ),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/main.rs::handle_pr", "src/main.rs", "handle_pr"),
-                    node(
-                        "src/main.rs::resolve_pr_base_sha",
-                        "src/main.rs",
-                        "resolve_pr_base_sha",
-                    ),
-                ],
-                edges: vec![Edge {
-                    from: "src/main.rs::handle_pr".to_string(),
-                    to: "src/main.rs::resolve_pr_base_sha".to_string(),
-                    is_cycle: false,
-                }],
-                roots: vec!["src/main.rs::handle_pr".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-2 changed symbols in 1 file
-
-- fn handle_pr (src/main.rs)
-  - fn resolve_pr_base_sha (src/main.rs)
-
-## Definitions
-
-### fn handle_pr (src/main.rs)
-
-```
-fn handle_pr(args: PrArgs) -> Result<()>
-```
-
-### fn resolve_pr_base_sha (src/main.rs)
-
-```
-fn resolve_pr_base_sha() -> Result<String>
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_order_definitions_in_true_pre_order_when_first_child_has_its_own_child() {
-        // A -> B, A -> C (B before C in edge order), B -> D. True pre-order
-        // DFS visits A, then descends fully into B's subtree (B, D) before
-        // moving on to C: A, B, D, C. A naive "append to order when a node
-        // is pushed onto the stack" (rather than when it is actually
-        // visited/popped) would instead produce A, C, B, D, because C gets
-        // pushed onto the stack right after A even though B is visited
-        // first — this test pins the correct DFS order down as the full
-        // rendered string so both the "Change graph" tree and "Definitions"
-        // order are asserted together.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![
-                    symbol("src/lib.rs::a", "a", SymbolKind::Function, "fn a()"),
-                    symbol("src/lib.rs::b", "b", SymbolKind::Function, "fn b()"),
-                    symbol("src/lib.rs::c", "c", SymbolKind::Function, "fn c()"),
-                    symbol("src/lib.rs::d", "d", SymbolKind::Function, "fn d()"),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::a", "src/lib.rs", "a"),
-                    node("src/lib.rs::b", "src/lib.rs", "b"),
-                    node("src/lib.rs::c", "src/lib.rs", "c"),
-                    node("src/lib.rs::d", "src/lib.rs", "d"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "src/lib.rs::a".to_string(),
-                        to: "src/lib.rs::b".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/lib.rs::a".to_string(),
-                        to: "src/lib.rs::c".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/lib.rs::b".to_string(),
-                        to: "src/lib.rs::d".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec!["src/lib.rs::a".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-4 changed symbols in 1 file
-
-- fn a (src/lib.rs)
-  - fn b (src/lib.rs)
-    - fn d (src/lib.rs)
-  - fn c (src/lib.rs)
-
-## Definitions
-
-### fn a (src/lib.rs)
-
-```
-fn a()
-```
-
-### fn b (src/lib.rs)
-
-```
-fn b()
-```
-
-### fn d (src/lib.rs)
-
-```
-fn d()
-```
-
-### fn c (src/lib.rs)
-
-```
-fn c()
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_mark_see_above_when_symbol_reachable_from_multiple_roots() {
-        // Both "foo" and "bar" reference "shared": it must be rendered in
-        // full once (under "foo", the first root in source order) and
-        // referenced by name only under "bar" (ADR 0008).
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![
-                    symbol("src/lib.rs::foo", "foo", SymbolKind::Function, "fn foo()"),
-                    symbol("src/lib.rs::bar", "bar", SymbolKind::Function, "fn bar()"),
-                    symbol(
-                        "src/lib.rs::shared",
-                        "shared",
-                        SymbolKind::Function,
-                        "fn shared()",
-                    ),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::foo", "src/lib.rs", "foo"),
-                    node("src/lib.rs::bar", "src/lib.rs", "bar"),
-                    node("src/lib.rs::shared", "src/lib.rs", "shared"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "src/lib.rs::foo".to_string(),
-                        to: "src/lib.rs::shared".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/lib.rs::bar".to_string(),
-                        to: "src/lib.rs::shared".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec!["src/lib.rs::foo".to_string(), "src/lib.rs::bar".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-3 changed symbols in 1 file
-
-- fn foo (src/lib.rs)
-  - fn shared (src/lib.rs)
-- fn bar (src/lib.rs)
-  - fn shared (src/lib.rs) (see above)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-### fn shared (src/lib.rs)
-
-```
-fn shared()
-```
-
-### fn bar (src/lib.rs)
-
-```
-fn bar()
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_cycle_warning_when_edge_is_marked_as_cycle() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/git.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/git.rs::resolve_pr_base_sha",
-                    "resolve_pr_base_sha",
-                    SymbolKind::Function,
-                    "fn resolve_pr_base_sha()",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node(
-                    "src/git.rs::resolve_pr_base_sha",
-                    "src/git.rs",
-                    "resolve_pr_base_sha",
-                )],
-                edges: vec![Edge {
-                    from: "src/git.rs::resolve_pr_base_sha".to_string(),
-                    to: "src/git.rs::resolve_pr_base_sha".to_string(),
-                    is_cycle: true,
-                }],
-                roots: vec!["src/git.rs::resolve_pr_base_sha".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn resolve_pr_base_sha (src/git.rs)
-  - ⚠️ fn resolve_pr_base_sha (src/git.rs) — dependency cycle, see above
-
-## Definitions
-
-### fn resolve_pr_base_sha (src/git.rs)
-
-```
-fn resolve_pr_base_sha()
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_full_cycle_example_with_two_root_functions_and_a_dependency_cycle() {
-        // The scenario from the ADR walkthrough: `handle_pr` calls
-        // `resolve_pr_base_sha`, which calls `fetch_base_branch` and also
-        // (a design smell the tool should surface) calls back into
-        // itself. `Config` is an unrelated, independent root.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![
-                FileReport {
-                    path: "src/main.rs".to_string(),
-                    symbols: vec![symbol(
-                        "src/main.rs::handle_pr",
-                        "handle_pr",
-                        SymbolKind::Function,
-                        "fn handle_pr(args: PrArgs) -> Result<()>",
-                    )],
-                },
-                FileReport {
-                    path: "src/git.rs".to_string(),
-                    symbols: vec![
-                        symbol(
-                            "src/git.rs::resolve_pr_base_sha",
-                            "resolve_pr_base_sha",
-                            SymbolKind::Function,
-                            "fn resolve_pr_base_sha() -> Result<String>",
-                        ),
-                        symbol(
-                            "src/git.rs::fetch_base_branch",
-                            "fetch_base_branch",
-                            SymbolKind::Function,
-                            "fn fetch_base_branch() -> Result<()>",
-                        ),
-                    ],
-                },
-                FileReport {
-                    path: "src/config.rs".to_string(),
-                    symbols: vec![symbol(
-                        "src/config.rs::Config",
-                        "Config",
-                        SymbolKind::Struct,
-                        "struct Config { path: String }",
-                    )],
-                },
-            ],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/main.rs::handle_pr", "src/main.rs", "handle_pr"),
-                    node(
-                        "src/git.rs::resolve_pr_base_sha",
-                        "src/git.rs",
-                        "resolve_pr_base_sha",
-                    ),
-                    node(
-                        "src/git.rs::fetch_base_branch",
-                        "src/git.rs",
-                        "fetch_base_branch",
-                    ),
-                    node("src/config.rs::Config", "src/config.rs", "Config"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "src/main.rs::handle_pr".to_string(),
-                        to: "src/git.rs::resolve_pr_base_sha".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/git.rs::resolve_pr_base_sha".to_string(),
-                        to: "src/git.rs::fetch_base_branch".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/git.rs::resolve_pr_base_sha".to_string(),
-                        to: "src/git.rs::resolve_pr_base_sha".to_string(),
-                        is_cycle: true,
-                    },
-                ],
-                roots: vec![
-                    "src/main.rs::handle_pr".to_string(),
-                    "src/config.rs::Config".to_string(),
-                ],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-4 changed symbols in 3 files — most in src/git.rs (2)
-
-- fn handle_pr (src/main.rs)
-  - fn resolve_pr_base_sha (src/git.rs)
-    - fn fetch_base_branch (src/git.rs)
-    - ⚠️ fn resolve_pr_base_sha (src/git.rs) — dependency cycle, see above
-- struct Config (src/config.rs)
-
-## Definitions
-
-### fn handle_pr (src/main.rs)
-
-```
-fn handle_pr(args: PrArgs) -> Result<()>
-```
-
-### fn resolve_pr_base_sha (src/git.rs)
-
-```
-fn resolve_pr_base_sha() -> Result<String>
-```
-
-### fn fetch_base_branch (src/git.rs)
-
-```
-fn fetch_base_branch() -> Result<()>
-```
-
-### struct Config (src/config.rs)
-
-```
-struct Config { path: String }
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_inline_two_leaf_struct_children_as_uses_annotation_on_method_line() {
-        // A method referencing two childless, non-function structs (the
-        // request/response shape the ADR calls out): both fold into the
-        // parent's own line as `— uses: ...` instead of rendering as their
-        // own nested lines, but both still get full "### ..." entries
-        // under "Definitions" (ADR 0012 decision 1).
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "store/items.go".to_string(),
-                symbols: vec![
-                    symbol(
-                        "store/items.go::UpsertItems",
-                        "UpsertItems",
-                        SymbolKind::Function,
-                        "func UpsertItems(req UpsertItemsRequest) (UpsertItemsResponse, error)",
-                    ),
-                    symbol(
-                        "store/items.go::UpsertItemsRequest",
-                        "UpsertItemsRequest",
-                        SymbolKind::Struct,
-                        "type UpsertItemsRequest struct { Items []Item }",
-                    ),
-                    symbol(
-                        "store/items.go::UpsertItemsResponse",
-                        "UpsertItemsResponse",
-                        SymbolKind::Struct,
-                        "type UpsertItemsResponse struct { Count int }",
-                    ),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node(
-                        "store/items.go::UpsertItems",
-                        "store/items.go",
-                        "UpsertItems",
-                    ),
-                    node(
-                        "store/items.go::UpsertItemsRequest",
-                        "store/items.go",
-                        "UpsertItemsRequest",
-                    ),
-                    node(
-                        "store/items.go::UpsertItemsResponse",
-                        "store/items.go",
-                        "UpsertItemsResponse",
-                    ),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "store/items.go::UpsertItems".to_string(),
-                        to: "store/items.go::UpsertItemsRequest".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "store/items.go::UpsertItems".to_string(),
-                        to: "store/items.go::UpsertItemsResponse".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec!["store/items.go::UpsertItems".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-3 changed symbols in 1 file
-
-- fn UpsertItems (store/items.go) — uses: UpsertItemsRequest, UpsertItemsResponse
-
-## Definitions
-
-### fn UpsertItems (store/items.go)
-
-```
-func UpsertItems(req UpsertItemsRequest) (UpsertItemsResponse, error)
-```
-
-### struct UpsertItemsRequest (store/items.go)
-
-```
-type UpsertItemsRequest struct { Items []Item }
-```
-
-### struct UpsertItemsResponse (store/items.go)
-
-```
-type UpsertItemsResponse struct { Count int }
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_disambiguate_folded_names_when_duplicate_symbols_fold_under_same_parent() {
-        // Two distinct `Dup` structs in the same file (mirroring an
-        // overloaded/shadowed-name scenario `graph::collect_nodes`
-        // disambiguates by appending `@{start_line}` to the node id) both
-        // fold under `foo`. Bare `Dup, Dup` would be ambiguous — Definitions
-        // shows two distinct `### struct Dup (src/lib.rs:5)` /
-        // `(src/lib.rs:10)` headers, so the folded annotation must use the
-        // same `Name (path:line)` form `tree_label` already uses for
-        // disambiguated symbols, not the bare name.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![
-                    symbol("src/lib.rs::foo", "foo", SymbolKind::Function, "fn foo()"),
-                    ExtractedSymbol {
-                        range: LineRange { start: 5, end: 6 },
-                        ..symbol(
-                            "src/lib.rs::Dup@5",
-                            "Dup",
-                            SymbolKind::Struct,
-                            "struct Dup { a: i32 }",
-                        )
-                    },
-                    ExtractedSymbol {
-                        range: LineRange { start: 10, end: 11 },
-                        ..symbol(
-                            "src/lib.rs::Dup@10",
-                            "Dup",
-                            SymbolKind::Struct,
-                            "struct Dup { b: i32 }",
-                        )
-                    },
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::foo", "src/lib.rs", "foo"),
-                    node("src/lib.rs::Dup@5", "src/lib.rs", "Dup"),
-                    node("src/lib.rs::Dup@10", "src/lib.rs", "Dup"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "src/lib.rs::foo".to_string(),
-                        to: "src/lib.rs::Dup@5".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/lib.rs::foo".to_string(),
-                        to: "src/lib.rs::Dup@10".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-3 changed symbols in 1 file
-
-- fn foo (src/lib.rs) — uses: Dup (src/lib.rs:5), Dup (src/lib.rs:10)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-### struct Dup (src/lib.rs:5)
-
-```
-struct Dup { a: i32 }
-```
-
-### struct Dup (src/lib.rs:10)
-
-```
-struct Dup { b: i32 }
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_repeat_folded_struct_annotation_on_every_referencing_parent() {
-        // Both `foo` and `bar` reference the same childless struct `Shared`
-        // — unlike function children (which get a single full render plus
-        // `(see above)` elsewhere, ADR 0008), a folded name has no
-        // "see above" tracking: it legitimately repeats verbatim in the
-        // `— uses: ...` annotation on every parent that references it, and
-        // it must never itself get a `(see above)` line.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![
-                    symbol("src/lib.rs::foo", "foo", SymbolKind::Function, "fn foo()"),
-                    symbol("src/lib.rs::bar", "bar", SymbolKind::Function, "fn bar()"),
-                    symbol(
-                        "src/lib.rs::Shared",
-                        "Shared",
-                        SymbolKind::Struct,
-                        "struct Shared { x: i32 }",
-                    ),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::foo", "src/lib.rs", "foo"),
-                    node("src/lib.rs::bar", "src/lib.rs", "bar"),
-                    node("src/lib.rs::Shared", "src/lib.rs", "Shared"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "src/lib.rs::foo".to_string(),
-                        to: "src/lib.rs::Shared".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/lib.rs::bar".to_string(),
-                        to: "src/lib.rs::Shared".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec!["src/lib.rs::foo".to_string(), "src/lib.rs::bar".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-3 changed symbols in 1 file
-
-- fn foo (src/lib.rs) — uses: Shared
-- fn bar (src/lib.rs) — uses: Shared
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-### struct Shared (src/lib.rs)
-
-```
-struct Shared { x: i32 }
-```
-
-### fn bar (src/lib.rs)
-
-```
-fn bar()
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_childless_non_function_root_as_top_level_line_when_it_would_otherwise_be_foldable()
-     {
-        // `Config` is a childless struct — foldable by the structural
-        // criterion — but it is also a root, so it must still render as
-        // its own top-level tree line rather than being folded away
-        // entirely (roots are always their own top-level DFS start).
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/config.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/config.rs::Config",
-                    "Config",
-                    SymbolKind::Struct,
-                    "struct Config { path: String }",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/config.rs::Config", "src/config.rs", "Config")],
-                edges: vec![],
-                roots: vec!["src/config.rs::Config".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- struct Config (src/config.rs)
-
-## Definitions
-
-### struct Config (src/config.rs)
-
-```
-struct Config { path: String }
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_nested_line_when_non_function_child_has_its_own_children() {
-        // `Wrapper` is a non-function child of `foo`, but it is not
-        // foldable because it has an outgoing edge of its own (to `Inner`)
-        // — the structural criterion is "childless", not "non-function",
-        // so `Wrapper` itself still renders as a nested line exactly as
-        // before this feature. `Inner`, in turn, *is* childless and
-        // non-function, so it folds into `Wrapper`'s own line instead of
-        // getting a third nesting level.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![
-                    symbol("src/lib.rs::foo", "foo", SymbolKind::Function, "fn foo()"),
-                    symbol(
-                        "src/lib.rs::Wrapper",
-                        "Wrapper",
-                        SymbolKind::Struct,
-                        "struct Wrapper { inner: Inner }",
-                    ),
-                    symbol(
-                        "src/lib.rs::Inner",
-                        "Inner",
-                        SymbolKind::Struct,
-                        "struct Inner { x: i32 }",
-                    ),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::foo", "src/lib.rs", "foo"),
-                    node("src/lib.rs::Wrapper", "src/lib.rs", "Wrapper"),
-                    node("src/lib.rs::Inner", "src/lib.rs", "Inner"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "src/lib.rs::foo".to_string(),
-                        to: "src/lib.rs::Wrapper".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/lib.rs::Wrapper".to_string(),
-                        to: "src/lib.rs::Inner".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-3 changed symbols in 1 file
-
-- fn foo (src/lib.rs)
-  - struct Wrapper (src/lib.rs) — uses: Inner
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-### struct Wrapper (src/lib.rs)
-
-```
-struct Wrapper { inner: Inner }
-```
-
-### struct Inner (src/lib.rs)
-
-```
-struct Inner { x: i32 }
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_not_fold_non_function_child_when_its_only_children_are_cycle_edges() {
-        // `Node` is a non-function type whose only outgoing edge is a
-        // cycle edge back to itself — `children_by_node` still records an
-        // entry for it, so it is *not* foldable (folding requires no
-        // outgoing edges at all) and must render as its own nested line
-        // with the cycle warning still visible beneath it.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![
-                    symbol("src/lib.rs::foo", "foo", SymbolKind::Function, "fn foo()"),
-                    symbol(
-                        "src/lib.rs::Node",
-                        "Node",
-                        SymbolKind::Struct,
-                        "struct Node { next: Option<Box<Node>> }",
-                    ),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("src/lib.rs::foo", "src/lib.rs", "foo"),
-                    node("src/lib.rs::Node", "src/lib.rs", "Node"),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "src/lib.rs::foo".to_string(),
-                        to: "src/lib.rs::Node".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "src/lib.rs::Node".to_string(),
-                        to: "src/lib.rs::Node".to_string(),
-                        is_cycle: true,
-                    },
-                ],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-2 changed symbols in 1 file
-
-- fn foo (src/lib.rs)
-  - struct Node (src/lib.rs)
-    - ⚠️ struct Node (src/lib.rs) — dependency cycle, see above
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-### struct Node (src/lib.rs)
-
-```
-struct Node { next: Option<Box<Node>> }
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_container_comment_when_symbol_has_container() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    container: Some("impl Foo".to_string()),
-                    ..symbol(
-                        "src/lib.rs::bar",
-                        "bar",
-                        SymbolKind::Function,
-                        "fn bar(&self) -> i32",
-                    )
-                }],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::bar", "src/lib.rs", "bar")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::bar".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn bar (src/lib.rs)
-
-## Definitions
-
-### fn bar (src/lib.rs)
-
-```
-// impl Foo
-fn bar(&self) -> i32
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_depends_on_list_when_symbol_has_dependencies() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    dependencies: vec![crate::deps::ResolvedSymbol {
-                        signature: "struct Point { x: i32, y: i32 }".to_string(),
-                        path: "src/point.rs".to_string(),
-                    }],
-                    ..symbol(
-                        "src/lib.rs::foo",
-                        "foo",
-                        SymbolKind::Function,
-                        "fn foo(p: Point) -> i32",
-                    )
-                }],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo(p: Point) -> i32
-```
-
-Depends on:
-- `src/point.rs`: `struct Point { x: i32, y: i32 }`
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_multiple_depends_on_entries_when_symbol_has_several_dependencies() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    dependencies: vec![
-                        crate::deps::ResolvedSymbol {
-                            signature: "struct Point { x: i32 }".to_string(),
-                            path: "src/a.rs".to_string(),
-                        },
-                        crate::deps::ResolvedSymbol {
-                            signature: "struct Point { y: i32 }".to_string(),
-                            path: "src/b.rs".to_string(),
-                        },
-                    ],
-                    ..symbol(
-                        "src/lib.rs::foo",
-                        "foo",
-                        SymbolKind::Function,
-                        "fn foo(p: Point) -> i32",
-                    )
-                }],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo(p: Point) -> i32
-```
-
-Depends on:
-- `src/a.rs`: `struct Point { x: i32 }`
-- `src/b.rs`: `struct Point { y: i32 }`
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_omitted_matches_note_when_dependency_matches_were_capped() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    dependencies: vec![crate::deps::ResolvedSymbol {
-                        signature: "struct Point { x: i32 }".to_string(),
-                        path: "src/a.rs".to_string(),
-                    }],
-                    omitted_dependency_matches: 2,
-                    ..symbol(
-                        "src/lib.rs::foo",
-                        "foo",
-                        SymbolKind::Function,
-                        "fn foo(p: Point) -> i32",
-                    )
-                }],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo(p: Point) -> i32
-```
-
-Depends on:
-- `src/a.rs`: `struct Point { x: i32 }`
-- (+2 more definitions matched by name)
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    // Regression test: a signature containing a backtick code fence (e.g. a
-    // doc comment example embedded in a macro invocation) used to break out
-    // of the surrounding Markdown fence because it was always rendered with
-    // exactly 3 backticks. The fence length must be at least one longer
-    // than the longest run of backticks appearing in the rendered content.
-    #[test]
-    fn should_widen_fence_when_signature_contains_a_backtick_run() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/lib.rs::example_macro",
-                    "example_macro",
-                    SymbolKind::Function,
-                    "fn example_macro() { let s = \"```rust\\nfn f() {}\\n```\"; }",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node(
-                    "src/lib.rs::example_macro",
-                    "src/lib.rs",
-                    "example_macro",
-                )],
-                edges: vec![],
-                roots: vec!["src/lib.rs::example_macro".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn example_macro (src/lib.rs)
-
-## Definitions
-
-### fn example_macro (src/lib.rs)
-
-````
-fn example_macro() { let s = \"```rust\\nfn f() {}\\n```\"; }
-````
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    // Regression test: the container comment is part of the fenced block
-    // too, so a backtick run inside the container name must also widen the
-    // fence.
-    #[test]
-    fn should_widen_fence_when_container_contains_a_backtick_run() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![ExtractedSymbol {
-                    container: Some("impl Foo /* ```` */".to_string()),
-                    ..symbol(
-                        "src/lib.rs::bar",
-                        "bar",
-                        SymbolKind::Function,
-                        "fn bar(&self) -> i32",
-                    )
-                }],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::bar", "src/lib.rs", "bar")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::bar".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn bar (src/lib.rs)
-
-## Definitions
-
-### fn bar (src/lib.rs)
-
-`````
-// impl Foo /* ```` */
-fn bar(&self) -> i32
-`````
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_skipped_files_section_when_report_has_skips() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![
-                SkippedFile {
-                    path: "assets/logo.png".to_string(),
-                    reason: SkipReason::Binary,
-                },
-                SkippedFile {
-                    path: "src/main.py".to_string(),
-                    reason: SkipReason::UnsupportedLanguage,
-                },
-                SkippedFile {
-                    path: "src/old.rs".to_string(),
-                    reason: SkipReason::Deleted,
-                },
-            ],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Skipped files
-
-- assets/logo.png (binary)
-- src/main.py (unsupported language)
-- src/old.rs (deleted)
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_change_graph_then_skipped_section_when_report_has_both() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/lib.rs::foo",
-                    "foo",
-                    SymbolKind::Function,
-                    "fn foo()",
-                )],
-            }],
-            skipped: vec![SkippedFile {
-                path: "assets/logo.png".to_string(),
-                reason: SkipReason::Binary,
-            }],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-## Skipped files
-
-- assets/logo.png (binary)
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_omit_hotspots_section_when_hotspots_is_empty() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/lib.rs::foo",
-                    "foo",
-                    SymbolKind::Function,
-                    "fn foo()",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_hotspots_section_between_change_graph_and_definitions_when_hotspots_is_non_empty()
-     {
-        // `UpsertItemsRequest` (a struct) is referenced by two changed
-        // functions — the label reuses tree_label's `{kind} {name}
-        // ({path})` form, so the line reads
-        // "struct UpsertItemsRequest (store/items.go) — used by 2: ..."
-        // exactly as the ADR spec requires, and used_by names are joined
-        // in the order `compute_hotspots` already sorted them in (not
-        // re-sorted here).
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "store/items.go".to_string(),
-                symbols: vec![
-                    symbol(
-                        "store/items.go::HandleFoo",
-                        "HandleFoo",
-                        SymbolKind::Function,
-                        "func HandleFoo(req UpsertItemsRequest) error",
-                    ),
-                    symbol(
-                        "store/items.go::HandleBar",
-                        "HandleBar",
-                        SymbolKind::Function,
-                        "func HandleBar(req UpsertItemsRequest) error",
-                    ),
-                    symbol(
-                        "store/items.go::UpsertItemsRequest",
-                        "UpsertItemsRequest",
-                        SymbolKind::Struct,
-                        "type UpsertItemsRequest struct { Items []Item }",
-                    ),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("store/items.go::HandleFoo", "store/items.go", "HandleFoo"),
-                    node("store/items.go::HandleBar", "store/items.go", "HandleBar"),
-                    node(
-                        "store/items.go::UpsertItemsRequest",
-                        "store/items.go",
-                        "UpsertItemsRequest",
-                    ),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "store/items.go::HandleFoo".to_string(),
-                        to: "store/items.go::UpsertItemsRequest".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "store/items.go::HandleBar".to_string(),
-                        to: "store/items.go::UpsertItemsRequest".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec![
-                    "store/items.go::HandleFoo".to_string(),
-                    "store/items.go::HandleBar".to_string(),
-                ],
-            },
-            tests: vec![],
-            hotspots: vec![Hotspot {
-                id: "store/items.go::UpsertItemsRequest".to_string(),
-                path: "store/items.go".to_string(),
-                name: "UpsertItemsRequest".to_string(),
-                used_by: vec!["HandleBar".to_string(), "HandleFoo".to_string()],
-            }],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-3 changed symbols in 1 file
-
-- fn HandleFoo (store/items.go) — uses: UpsertItemsRequest
-- fn HandleBar (store/items.go) — uses: UpsertItemsRequest
-
-## Hotspots
-
-- struct UpsertItemsRequest (store/items.go) — used by 2: HandleBar, HandleFoo
-
-## Definitions
-
-### fn HandleFoo (store/items.go)
-
-```
-func HandleFoo(req UpsertItemsRequest) error
-```
-
-### struct UpsertItemsRequest (store/items.go)
-
-```
-type UpsertItemsRequest struct { Items []Item }
-```
-
-### fn HandleBar (store/items.go)
-
-```
-func HandleBar(req UpsertItemsRequest) error
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_hotspot_line_for_symbol_with_no_matching_definition() {
-        // Same defensive rationale as the "NOTE" block below: `hotspots`
-        // could in principle reference a node id with no corresponding
-        // `ExtractedSymbol` in `files` (the node is present in `graph`, so
-        // the empty-output guard does not short-circuit, but `files` itself
-        // has no matching symbol — mirroring the other lookup-miss tests'
-        // setup). Unlike "Change graph"/"Definitions" (which use
-        // `SymbolLookup`, keyed by symbol id, to find the container/
-        // signature and skip the line entirely on a miss), the "Hotspots"
-        // line still renders on a lookup miss, falling back to a bare
-        // `{name} ({path})` label with no kind prefix, rather than being
-        // dropped outright.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::ghost", "src/lib.rs", "ghost")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::ghost".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![Hotspot {
-                id: "src/lib.rs::ghost".to_string(),
-                path: "src/lib.rs".to_string(),
-                name: "ghost".to_string(),
-                used_by: vec!["a".to_string(), "b".to_string()],
-            }],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-
-## Hotspots
-
-- ghost (src/lib.rs) — used by 2: a, b
-
-## Definitions
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_file_size_warnings_section_when_warnings_are_present() {
-        // Two warnings across both severities: the `Split` entry must come
-        // first (ADR 0028 orders `Split` before `Warn`), each glyph and the
-        // threshold-numbered explanation pinned exactly as the ADR spec
-        // shows.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/lib.rs::foo",
-                    "foo",
-                    SymbolKind::Function,
-                    "fn foo()",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![
-                crate::file_size::FileSizeWarning {
-                    path: "b.rs".to_string(),
-                    line_count: 2500,
-                    severity: crate::file_size::FileSizeSeverity::Split,
-                },
-                crate::file_size::FileSizeWarning {
-                    path: "a.rs".to_string(),
-                    line_count: 1600,
-                    severity: crate::file_size::FileSizeSeverity::Warn,
-                },
-            ],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## File size warnings
-
-- 🚨 `b.rs` (2500 lines) — over the 2000-line split threshold
-- ⚠ `a.rs` (1600 lines) — over the 1500-line watch threshold; consider splitting
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_omit_file_size_warnings_section_when_report_has_no_warnings() {
-        // Empty `file_size_warnings` must drop the whole section — no bare
-        // "## File size warnings" heading with nothing under it.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/lib.rs::foo",
-                    "foo",
-                    SymbolKind::Function,
-                    "fn foo()",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert!(!actual.contains("## File size warnings"));
-    }
-
-    #[test]
-    fn should_place_file_size_warnings_between_hotspots_and_definitions() {
-        // A report with both a hotspot and a file-size warning: the
-        // `## File size warnings` section must land after `## Hotspots` and
-        // before `## Definitions`, matching ADR 0028's placement decision.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "store/items.go".to_string(),
-                symbols: vec![
-                    symbol(
-                        "store/items.go::HandleFoo",
-                        "HandleFoo",
-                        SymbolKind::Function,
-                        "func HandleFoo(req UpsertItemsRequest) error",
-                    ),
-                    symbol(
-                        "store/items.go::HandleBar",
-                        "HandleBar",
-                        SymbolKind::Function,
-                        "func HandleBar(req UpsertItemsRequest) error",
-                    ),
-                    symbol(
-                        "store/items.go::UpsertItemsRequest",
-                        "UpsertItemsRequest",
-                        SymbolKind::Struct,
-                        "type UpsertItemsRequest struct { Items []Item }",
-                    ),
-                ],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![
-                    node("store/items.go::HandleFoo", "store/items.go", "HandleFoo"),
-                    node("store/items.go::HandleBar", "store/items.go", "HandleBar"),
-                    node(
-                        "store/items.go::UpsertItemsRequest",
-                        "store/items.go",
-                        "UpsertItemsRequest",
-                    ),
-                ],
-                edges: vec![
-                    Edge {
-                        from: "store/items.go::HandleFoo".to_string(),
-                        to: "store/items.go::UpsertItemsRequest".to_string(),
-                        is_cycle: false,
-                    },
-                    Edge {
-                        from: "store/items.go::HandleBar".to_string(),
-                        to: "store/items.go::UpsertItemsRequest".to_string(),
-                        is_cycle: false,
-                    },
-                ],
-                roots: vec![
-                    "store/items.go::HandleFoo".to_string(),
-                    "store/items.go::HandleBar".to_string(),
-                ],
-            },
-            tests: vec![],
-            hotspots: vec![Hotspot {
-                id: "store/items.go::UpsertItemsRequest".to_string(),
-                path: "store/items.go".to_string(),
-                name: "UpsertItemsRequest".to_string(),
-                used_by: vec!["HandleBar".to_string(), "HandleFoo".to_string()],
-            }],
-            file_size_warnings: vec![crate::file_size::FileSizeWarning {
-                path: "store/items.go".to_string(),
-                line_count: 2100,
-                severity: crate::file_size::FileSizeSeverity::Split,
-            }],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-3 changed symbols in 1 file
-
-- fn HandleFoo (store/items.go) — uses: UpsertItemsRequest
-- fn HandleBar (store/items.go) — uses: UpsertItemsRequest
-
-## Hotspots
-
-- struct UpsertItemsRequest (store/items.go) — used by 2: HandleBar, HandleFoo
-
-## File size warnings
-
-- 🚨 `store/items.go` (2100 lines) — over the 2000-line split threshold
-
-## Definitions
-
-### fn HandleFoo (store/items.go)
-
-```
-func HandleFoo(req UpsertItemsRequest) error
-```
-
-### struct UpsertItemsRequest (store/items.go)
-
-```
-type UpsertItemsRequest struct { Items []Item }
-```
-
-### fn HandleBar (store/items.go)
-
-```
-func HandleBar(req UpsertItemsRequest) error
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    // ADR 0028: JSON output must always carry `file_size_warnings` as a
-    // top-level field, present-and-empty when nothing warns (mirroring how
-    // `hotspots` is always present). This pins that shape end-to-end via
-    // the same `render` entry point the Markdown tests above use.
-    #[test]
-    fn should_include_file_size_warnings_in_json_output() {
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![],
-                edges: vec![],
-                roots: vec![],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![
-                crate::file_size::FileSizeWarning {
-                    path: "b.rs".to_string(),
-                    line_count: 2500,
-                    severity: crate::file_size::FileSizeSeverity::Split,
-                },
-                crate::file_size::FileSizeWarning {
-                    path: "a.rs".to_string(),
-                    line_count: 1600,
-                    severity: crate::file_size::FileSizeSeverity::Warn,
-                },
-            ],
-            removed: vec![],
-        };
-
-        let expected = "\
-{
-  \"files\": [],
-  \"skipped\": [],
-  \"graph\": {
-    \"nodes\": [],
-    \"edges\": [],
-    \"roots\": []
-  },
-  \"tests\": [],
-  \"hotspots\": [],
-  \"file_size_warnings\": [
-    {
-      \"path\": \"b.rs\",
-      \"line_count\": 2500,
-      \"severity\": \"split\"
-    },
-    {
-      \"path\": \"a.rs\",
-      \"line_count\": 1600,
-      \"severity\": \"warn\"
-    }
-  ],
-  \"removed\": []
-}"
-        .to_string();
-        let actual = render(&report, OutputFormat::Json).expect("json render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    // NOTE: these three tests hand-build a `Report` whose `graph` refers to
-    // node ids that have no corresponding `ExtractedSymbol` in `files` — an
-    // inconsistency `pipeline::analyze_diff` never actually produces (the
-    // graph is always built from, and ids stamped onto, the very same
-    // `files` list), but exercised here defensively so the lookup-miss
-    // fallback branches (`SymbolLookup::get` returning `None`) have direct
-    // coverage rather than being unreachable-in-practice dead code.
-
-    #[test]
-    fn should_skip_definitions_entry_when_visit_order_id_has_no_matching_symbol() {
-        // `dfs_pre_order`'s `visit_order` is derived from `graph.nodes`, not
-        // `files`, so a node with no matching `ExtractedSymbol` reaches the
-        // `let Some(..) = lookup.get(id) else { continue }` branch in the
-        // "Definitions" loop; the malformed root must simply be skipped
-        // rather than panicking or emitting a broken heading.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::ghost", "src/lib.rs", "ghost")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::ghost".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-
-## Definitions
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_render_nothing_for_root_when_root_id_has_no_matching_symbol() {
-        // Same lookup miss as above, but hit inside `render_tree_node`'s
-        // own `let Some(..) = lookup.get(id) else { return Ok(()) }` guard
-        // (the "Change graph" tree-line branch) rather than the
-        // "Definitions" loop.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::ghost", "src/lib.rs", "ghost")],
-                edges: vec![],
-                roots: vec!["src/lib.rs::ghost".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-
-## Definitions
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        // Both malformed-root branches (the "Change graph" line and the
-        // "Definitions" entry) are exercised by the same minimal report;
-        // asserted together since there is no simpler input that isolates
-        // only one of the two lookups.
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn should_omit_cycle_warning_line_when_cycle_target_id_has_no_matching_symbol() {
-        // A cycle edge whose `to` id has no matching symbol hits
-        // `render_tree_node`'s inner `let Some(..) = lookup.get(child_id)
-        // else { continue }` guard (the cycle-warning branch specifically,
-        // as opposed to the two tests above which exercise the
-        // non-cycle-edge lookups) — the warning line is simply omitted
-        // rather than rendering a broken label.
-        let report = Report {
-            origin: ReportOrigin::Diff,
-            files: vec![FileReport {
-                path: "src/lib.rs".to_string(),
-                symbols: vec![symbol(
-                    "src/lib.rs::foo",
-                    "foo",
-                    SymbolKind::Function,
-                    "fn foo()",
-                )],
-            }],
-            skipped: vec![],
-            graph: SymbolGraph {
-                nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                edges: vec![Edge {
-                    from: "src/lib.rs::foo".to_string(),
-                    to: "src/lib.rs::ghost".to_string(),
-                    is_cycle: true,
-                }],
-                roots: vec!["src/lib.rs::foo".to_string()],
-            },
-            tests: vec![],
-            hotspots: vec![],
-            file_size_warnings: vec![],
-            removed: vec![],
-        };
-
-        let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-"
-        .to_string();
-        let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-        assert_eq!(expected, actual);
-    }
-
-    mod classification_rendering_tests {
-        use super::*;
-        use crate::extract::{Classification, RemovedSymbol};
-        use pretty_assertions::assert_eq;
-        use rstest::rstest;
-
-        #[test]
-        fn should_append_new_marker_to_tree_and_definition_when_symbol_is_added() {
-            let mut foo = symbol("src/lib.rs::foo", "foo", SymbolKind::Function, "fn foo()");
-            foo.classification = Some(Classification::Added);
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![foo],
-                }],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                    edges: vec![],
-                    roots: vec!["src/lib.rs::foo".to_string()],
-                },
-                tests: vec![],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![],
-            };
-
-            let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs) — new
-
-## Definitions
-
-### fn foo (src/lib.rs) — new
-
-```
-fn foo()
-```
-
-"
-            .to_string();
-            let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert_eq!(expected, actual);
-        }
-
-        #[test]
-        fn should_render_diff_block_and_marker_when_symbol_is_signature_changed() {
-            let mut foo = symbol(
-                "src/lib.rs::foo",
-                "foo",
-                SymbolKind::Function,
-                "fn foo(a: i32, b: i32) -> i32",
-            );
-            foo.classification = Some(Classification::SignatureChanged);
-            foo.previous_signature = Some("fn foo(a: i32) -> i32".to_string());
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![foo],
-                }],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                    edges: vec![],
-                    roots: vec!["src/lib.rs::foo".to_string()],
-                },
-                tests: vec![],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![],
-            };
-
-            let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs) — signature changed
-
-## Definitions
-
-### fn foo (src/lib.rs) — signature changed
-
-```diff
--fn foo(a: i32) -> i32
-+fn foo(a: i32, b: i32) -> i32
-```
-
-"
-            .to_string();
-            let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert_eq!(expected, actual);
-        }
-
-        // A signature-changed symbol with a container gets the container
-        // comment line rendered unchanged above the diff lines — it is not
-        // itself part of the base/head comparison (see `render_definition`'s
-        // doc comment).
-        #[test]
-        fn should_render_container_comment_above_diff_lines_when_signature_changed_symbol_has_container()
-         {
-            let mut bar = symbol(
-                "src/lib.rs::bar",
-                "bar",
-                SymbolKind::Function,
-                "fn bar(&self, extra: i32) -> i32",
-            );
-            bar.container = Some("impl Foo".to_string());
-            bar.classification = Some(Classification::SignatureChanged);
-            bar.previous_signature = Some("fn bar(&self) -> i32".to_string());
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![bar],
-                }],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![node("src/lib.rs::bar", "src/lib.rs", "bar")],
-                    edges: vec![],
-                    roots: vec!["src/lib.rs::bar".to_string()],
-                },
-                tests: vec![],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![],
-            };
-
-            let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn bar (src/lib.rs) — signature changed
-
-## Definitions
-
-### fn bar (src/lib.rs) — signature changed
-
-```diff
-// impl Foo
--fn bar(&self) -> i32
-+fn bar(&self, extra: i32) -> i32
-```
-
-"
-            .to_string();
-            let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert_eq!(expected, actual);
-        }
-
-        // `body_only` and unattempted (`None`) classification both render
-        // completely unmarked — no `— <marker>` suffix anywhere, and a
-        // plain (non-diff) fenced signature block.
-        #[rstest]
-        #[case::should_render_unmarked_when_classification_is_body_only(Some(
-            Classification::BodyOnly
-        ))]
-        #[case::should_render_unmarked_when_classification_is_none(None)]
-        fn should_render_unmarked_tree_and_definition(
-            #[case] classification: Option<Classification>,
-        ) {
-            let mut foo = symbol("src/lib.rs::foo", "foo", SymbolKind::Function, "fn foo()");
-            foo.classification = classification;
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![foo],
-                }],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                    edges: vec![],
-                    roots: vec!["src/lib.rs::foo".to_string()],
-                },
-                tests: vec![],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![],
-            };
-
-            let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-"
-            .to_string();
-            let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert_eq!(expected, actual);
-        }
-
-        #[test]
-        fn should_append_marker_to_hotspot_line_before_used_by() {
-            let mut shared = symbol(
-                "src/lib.rs::shared",
-                "shared",
-                SymbolKind::Function,
-                "fn shared()",
-            );
-            shared.classification = Some(Classification::SignatureChanged);
-            shared.previous_signature = Some("fn shared(a: i32)".to_string());
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![shared],
-                }],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![node("src/lib.rs::shared", "src/lib.rs", "shared")],
-                    edges: vec![],
-                    roots: vec!["src/lib.rs::shared".to_string()],
-                },
-                tests: vec![],
-                hotspots: vec![Hotspot {
-                    id: "src/lib.rs::shared".to_string(),
-                    path: "src/lib.rs".to_string(),
-                    name: "shared".to_string(),
-                    used_by: vec!["caller_one".to_string(), "caller_two".to_string()],
-                }],
-                file_size_warnings: vec![],
-                removed: vec![],
-            };
-
-            let markdown =
-                render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-            // NOTE: partial assert (searching for one line) rather than a
-            // fully qualified comparison of the whole render — this test's
-            // concern is solely the "Hotspots" line's marker placement, and
-            // the "Change graph"/"Definitions" sections above it are
-            // already covered by other tests in this module (e.g.
-            // `should_render_diff_block_and_marker_when_symbol_is_signature_changed`).
-            let hotspots_line = markdown
-                .lines()
-                .find(|line| line.contains("used by"))
-                .expect("hotspots section must contain the shared symbol's line");
-
-            assert_eq!(
-                "- fn shared (src/lib.rs) — signature changed — used by 2: caller_one, caller_two",
-                hotspots_line
-            );
-        }
-
-        #[test]
-        fn should_render_removed_symbols_section_between_definitions_and_tests() {
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![symbol(
-                        "src/lib.rs::foo",
-                        "foo",
-                        SymbolKind::Function,
-                        "fn foo()",
-                    )],
-                }],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                    edges: vec![],
-                    roots: vec!["src/lib.rs::foo".to_string()],
-                },
-                tests: vec![TestFileSummary {
-                    path: "src/lib.rs".to_string(),
-                    symbol_count: 1,
-                }],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![RemovedSymbol {
-                    name: "old_helper".to_string(),
-                    kind: SymbolKind::Function,
-                    path: "src/lib.rs".to_string(),
-                    signature: "fn old_helper()".to_string(),
-                }],
-            };
-
-            let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs)
-
-## Definitions
-
-### fn foo (src/lib.rs)
-
-```
-fn foo()
-```
-
-## Removed symbols
-
-- fn old_helper (src/lib.rs)
-
-## Tests
-
-- src/lib.rs: 1 changed test symbol
-
-"
-            .to_string();
-            let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert_eq!(expected, actual);
-        }
-
-        // A diff whose only changed-symbol-level content is a removal (no
-        // graph nodes at all, e.g. a whole function deleted with nothing
-        // added back — see `pipeline::tests::classification_wiring_tests`'s
-        // "hunk only removes lines" case) must still render "## Removed
-        // symbols" on its own — the empty-output guard at the top of
-        // `render_markdown` must not treat an empty `graph.nodes` as "there
-        // is nothing to say" when `removed` is non-empty.
-        #[test]
-        fn should_render_removed_symbols_section_alone_when_graph_is_empty() {
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![],
-                    edges: vec![],
-                    roots: vec![],
-                },
-                tests: vec![],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![RemovedSymbol {
-                    name: "old_helper".to_string(),
-                    kind: SymbolKind::Function,
-                    path: "src/lib.rs".to_string(),
-                    signature: "fn old_helper()".to_string(),
-                }],
-            };
-
-            let expected = "\
-## Removed symbols
-
-- fn old_helper (src/lib.rs)
-
-"
-            .to_string();
-            let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert_eq!(expected, actual);
-        }
-
-        // Regression test: `removed_symbol_label` deliberately omits
-        // `container` (see its own doc comment), so two distinct removed
-        // symbols sharing name+kind but differing only by container (e.g.
-        // the same method name removed from two different impls in one
-        // file) render identical lines. Without deduplication this would
-        // print the same line twice; the section must show it once, in
-        // first-occurrence order.
-        #[test]
-        fn should_deduplicate_identical_removed_symbol_lines() {
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![],
-                    edges: vec![],
-                    roots: vec![],
-                },
-                tests: vec![],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![
-                    RemovedSymbol {
-                        name: "save".to_string(),
-                        kind: SymbolKind::Function,
-                        path: "src/lib.rs".to_string(),
-                        signature: "fn save(&self)".to_string(),
-                    },
-                    RemovedSymbol {
-                        name: "other".to_string(),
-                        kind: SymbolKind::Function,
-                        path: "src/lib.rs".to_string(),
-                        signature: "fn other(&self)".to_string(),
-                    },
-                    // Same name+kind+path as the first entry, but a
-                    // different container/signature — the label is
-                    // identical to the first line even though this is a
-                    // genuinely distinct removed symbol (different impl).
-                    RemovedSymbol {
-                        name: "save".to_string(),
-                        kind: SymbolKind::Function,
-                        path: "src/lib.rs".to_string(),
-                        signature: "fn save(&self, id: &str)".to_string(),
-                    },
-                ],
-            };
-
-            let expected = "\
-## Removed symbols
-
-- fn save (src/lib.rs)
-- fn other (src/lib.rs)
-
-"
-            .to_string();
-            let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert_eq!(expected, actual);
-        }
-
-        #[test]
-        fn should_omit_removed_symbols_section_when_removed_is_empty() {
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![symbol(
-                        "src/lib.rs::foo",
-                        "foo",
-                        SymbolKind::Function,
-                        "fn foo()",
-                    )],
-                }],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                    edges: vec![],
-                    roots: vec!["src/lib.rs::foo".to_string()],
-                },
-                tests: vec![],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![],
-            };
-
-            let markdown =
-                render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert!(!markdown.contains("## Removed symbols"));
-        }
-
-        // A diff isn't valid Markdown to embed unfenced — ensure the fence
-        // still widens against a backtick run appearing in either the base
-        // or head signature text, not just the head signature the way
-        // `fence_for` alone would check.
-        #[test]
-        fn should_widen_fence_when_previous_signature_contains_a_backtick_run() {
-            let mut foo = symbol(
-                "src/lib.rs::foo",
-                "foo",
-                SymbolKind::Function,
-                "fn foo() -> i32",
-            );
-            foo.classification = Some(Classification::SignatureChanged);
-            // Three consecutive backticks in the *base* signature only —
-            // proves the fence widens against `previous_signature`, not
-            // just the head `signature` the way plain `fence_for` would.
-            foo.previous_signature =
-                Some("fn foo() { let s = \"```rust\\nfn f() {}\\n```\"; }".to_string());
-            let report = Report {
-                origin: ReportOrigin::Diff,
-                files: vec![FileReport {
-                    path: "src/lib.rs".to_string(),
-                    symbols: vec![foo],
-                }],
-                skipped: vec![],
-                graph: SymbolGraph {
-                    nodes: vec![node("src/lib.rs::foo", "src/lib.rs", "foo")],
-                    edges: vec![],
-                    roots: vec!["src/lib.rs::foo".to_string()],
-                },
-                tests: vec![],
-                hotspots: vec![],
-                file_size_warnings: vec![],
-                removed: vec![],
-            };
-
-            let expected = "\
-## Change graph
-
-1 changed symbol in 1 file
-
-- fn foo (src/lib.rs) — signature changed
-
-## Definitions
-
-### fn foo (src/lib.rs) — signature changed
-
-````diff
--fn foo() { let s = \"```rust\\nfn f() {}\\n```\"; }
-+fn foo() -> i32
-````
-
-"
-            .to_string();
-            let actual = render(&report, OutputFormat::Markdown).expect("markdown render succeeds");
-
-            assert_eq!(expected, actual);
-        }
-    }
-}
+#[path = "markdown_tests/mod.rs"]
+mod tests;
